requirements.txt
fitz==0.0.1.dev2
numpy==1.26.4
pymongo==4.6.1
python-dotenv==1.0.1
replicate==0.23.1
scipy==1.12.0
sentence_transformers==2.3.1
streamlit==1.31.1
torch==2.2.0